backend/requirements.txt
Flask==2.3.2
Flask-Cors==3.0.10
requests==2.31.0
Pillow==10.0.0
SQLAlchemy==2.0.20
python-decouple==3.8
python-dotenv==1.1.0